rename: alert-modal 파일명 수정 및 폴더이동
