+++ components/custom/index.ts
@@ -2,3 +2,4 @@ export { default as CustomInput } from "./input";
 export { default as CustomCheckBox } from "./check-box";
 export { default as CustomNicknameEditForm } from "./nickname-edit-form";
 export { default as CustomSelectBox } from "./select-box";
+export { default as AlertModal } from "./alert-modal"
