+++ pages/profile/edit.tsx
@@ -3,7 +3,7 @@ import style from "styles/edit.module.css";
 import Image from "next/image";
 import { ICONS } from "lib/assets";
 import NicknameEditForm from "components/custom/nickname-edit-form";
-import AlertModal from "components/alertModal";
+import AlertModal from "components/custom/alert-modal";
 import useModal from "hooks/useModal";
 import EditIcon from "components/icons/edit";
